Rename job to match the class name
