+++ lib/capsulecrm/send_directory_entry_to_capsule.rb
