+++ lib/open-orgn-services.rb
@@ -29,7 +29,7 @@ require 'signup/signup_processor'
 require 'capsulecrm/capsule_helper'
 require 'capsulecrm/partner_enquiry_processor'
 require 'capsulecrm/send_signup_to_capsule'
-require 'capsulecrm/directory_entry_processor'
+require 'capsulecrm/send_directory_entry_to_capsule'
 require 'capsulecrm/capsule_sync_monitor'
 require 'capsulecrm/sync_capsule_data'
 require 'capsulecrm/save_membership_id_in_capsule'
